requirements.txt
pillow==10.2.0
python-dotenv==1.0.1
